Core Layer and Wrapper Layer are added to project.
The architecture of this project includes 3 layers.
The deepest layer is sokobanCore which contains the algorithms to solve sokoban game. this layer is written in c++.
The second one is Wrapper which builds dll from core layer to be able to use core layer's classes and methods in c#. This layer is written in cli/c++.
And the last one is ui layer which is written in c#.